rename constants to translations
