+++ x-pack/plugins/cloud_security_posture/public/pages/findings/findings.test.tsx
@@ -6,14 +6,14 @@
  */
 import React from 'react';
 import { Findings } from './findings';
-import { MISSING_KUBEBEAT } from './constants';
+import { MISSING_KUBEBEAT } from './translations';
 import { render, screen } from '@testing-library/react';
 import { TestProvider } from '../../application/test_provider';
 import { dataPluginMock } from '../../../../../../src/plugins/data/public/mocks';
 import { coreMock } from '../../../../../../src/core/public/mocks';
 import { createStubDataView } from '../../../../../../src/plugins/data_views/public/data_views/data_view.stub';
 import * as utils from './utils';
-import { CSP_KUBEBEAT_INDEX_NAME } from '../../../common/constants';
+import { CSP_KUBEBEAT_INDEX_NAME } from '../../../common/translations';
 import * as TEST_SUBJECTS from './test_subjects';
 import type { UseQueryResult } from 'react-query';
 import type { DataView } from '../../../../../../src/plugins/data/common';
