+++ x-pack/plugins/cloud_security_posture/public/pages/findings/translations.ts
