+++ x-pack/plugins/cloud_security_posture/public/pages/findings/utils.tsx
@@ -16,7 +16,7 @@ import type {
   TimeRange,
 } from '../../../../../../src/plugins/data/common';
 import type { CspPluginSetup } from '../../types';
-import { CSP_KUBEBEAT_INDEX_NAME } from '../../../common/constants';
+import { CSP_KUBEBEAT_INDEX_NAME } from '../../../common/translations';
 import { useKibana } from '../../../../../../src/plugins/kibana_react/public';
 
 export const extractErrorMessage = (e: unknown): string =>
